Rename `test` module to `debug`
