+++ backend/src/api/debug.rs
